MarketInsightAgencyy/requirements.txt
agency-swarm
pandas
spacy
statsmodels
matplotlib
jinja2
python-dotenv
gradio
flask
flask-socketio
flask-cors
psutil
deepgram-sdk
eventlet>=0.30.0 